+++ tests/Fixture/Container/EmptyContainerWithNamespace.php
@@ -1,22 +1,19 @@
 <?php
+
+declare(strict_types=1);
+
 namespace WoohooLabs\Zen\Tests\Fixture\Container;
 
 use WoohooLabs\Zen\AbstractCompiledContainer;
 
 class EmptyContainerWithNamespace extends AbstractCompiledContainer
 {
-    /**
-     * @var string[]
-     */
-    protected static $entryPoints = [
+    /** @var string[] */
+    protected static array $entryPoints = [
     ];
+    protected string $rootDirectory;
 
-    /**
-     * @var string
-     */
-    protected $rootDirectory;
-
-    public function __construct(string $rootDirectory = '')
+    public function __construct(string $rootDirectory = "")
     {
         $this->rootDirectory = $rootDirectory;
     }