+++ tests/Fixture/Container/EmptyContainerWithoutNamespace.php
@@ -1,21 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 use WoohooLabs\Zen\AbstractCompiledContainer;
 
 class EmptyContainerWithoutNamespace extends AbstractCompiledContainer
 {
-    /**
-     * @var string[]
-     */
-    protected static $entryPoints = [
+    /** @var string[] */
+    protected static array $entryPoints = [
     ];
+    protected string $rootDirectory;
 
-    /**
-     * @var string
-     */
-    protected $rootDirectory;
-
-    public function __construct(string $rootDirectory = '')
+    public function __construct(string $rootDirectory = "")
     {
         $this->rootDirectory = $rootDirectory;
     }